requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# UI and API
streamlit>=1.30.0
fastapi>=0.109.0
uvicorn>=0.27.0
